refactor: move linter to header
